% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_cndbias.R
\name{cndbias}
\alias{cndbias}
\title{Conditional bias of group sequential design}
\usage{
cndbias(i, t_vec, spd, alpha = 0.05, mu)
}
\arguments{
\item{i}{Index of interim look, e.g., \code{i=2} for the second interim look.
\code{i=length(t_vec)} for the final analysis.}

\item{t_vec}{Timing vector of interim looks and final analysis in information fraction. For example,
a trial design of one interim look at half information and final analysis has \code{t_vec=c(0.5,1)}}

\item{spd}{1 and 2 correspond to alpha spending functions which give O'Brien Fleming and Pocock
type boundaries, respectively. A value of 3 is the power family. See below for details.}

\item{alpha}{Type I error, default is 0.05.}

\item{mu}{drift parameter}
}
\value{
Conditional bias in drift scale given the trial has stopped at interim analysis of information
fraction \code{t_vec[i]}.
}
\description{
Calculate the conditional bias given that a trial stopped at an interim look
}
\section{Details}{
 Argument \code{spd} essentially utilizes argument \code{iuse} of function \code{bounds}
from package "\code{ldbounds}". Please refer package "\code{ldbounds}" for more details.
}

\examples{
#The conditional bias at the first interim look for a trial with 4 interim looks at information
#fraction 0.2,0.4,0.6 and 0.8, as well as with Pocock type boundary, and drift 5
> cndbias(i = 1,t_vec = c(0.2,0.4,0.6,0.8,1),spd = 2,alpha = 0.05,mu = 5)
[1] 2.081086
}
