% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_drift2hr.R
\name{drift2hr}
\alias{drift2hr}
\title{Transform from drift scale to hazard ratio scale}
\usage{
drift2hr(drift, trtfrac, maxevents)
}
\arguments{
\item{drift}{expectation of drift estimate}

\item{trtfrac}{treatment randomization fraction, e.g., \code{trtfrac} = 0.5 for balanced randomization}

\item{maxevents}{maximum event number}
}
\value{
lower bound of expectation of hazard ratio estimate corresponding to drift expectation, treatment randomization fraction and
maximum event number.
}
\description{
Calculate lower bound of expectation of hazard ratio estimate
}
