% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gen_bstar_vec.R
\name{gen_bstar_vec}
\alias{gen_bstar_vec}
\title{Lower boundary for \eqn{W(T)-\mu}T}}
\usage{
gen_bstar_vec(t_vec, c_vec, mu)
}
\arguments{
\item{t_vec}{Timing vector of interim looks and final analysis in information fraction. For example,
a trial design of one interim look at half information and final analysis has \code{t_vec=c(0.5,1)}}

\item{c_vec}{Symmetric boundaries for \eqn{W(T)/(T)^{1/2}}}

\item{mu}{drift parameter}
}
\value{
Lower boundaries vector with same length as \code{c_vec}
}
\description{
Lower boundary for \eqn{W(T)-\mu}T}
}
