% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gen_cstar_vec.R
\name{gen_cstar_vec}
\alias{gen_cstar_vec}
\title{Upper boundaries for \eqn{W(T)-\mu T}}
\usage{
gen_cstar_vec(t_vec, c_vec, mu)
}
\arguments{
\item{t_vec}{timing vector of interim looks and final analysis in information fraction. For example,
a trial design of one interim look at half information and final analysis has \code{t_vec=c(0.5,1)}}

\item{c_vec}{vector of symmetric boundaries for \eqn{W(T)/(T)^{1/2}}}

\item{mu}{drift parameter}
}
\value{
vector of upper boundaries for \eqn{W(T)-\mu T}
}
\description{
Calculate vector of upper boundaries for \eqn{W(T)-\mu T}
}
