% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_gen_sigma_vec.R
\name{gen_sigma_vec}
\alias{gen_sigma_vec}
\title{Calculate \eqn{\sigma}}
\usage{
gen_sigma_vec(t_vec)
}
\arguments{
\item{t_vec}{timing vector of interim looks and final analysis in information fraction. For example,
a trial design of one interim look at half information and final analysis has \code{t_vec=c(0.5,1)}}
}
\value{
vector of \eqn{\sigmai}, same length as \code{t_vec}.
}
\description{
Calculate \eqn{\sigma}
}
