% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_hr2drift.R
\name{hr2drift}
\alias{hr2drift}
\title{Transform from hazard ratio scale to drift scale}
\usage{
hr2drift(hr, trtfrac, maxevents)
}
\arguments{
\item{hr}{hazard ratio}

\item{trtfrac}{treatment randomization fraction, e.g., \code{trtfrac} = 0.5 for balanced randomization}

\item{maxevents}{maximum event number}
}
\value{
drift value corresponding to associated hazard ratio, treatment randomization fraction and
maximum event number.
}
\description{
Transform from hazard ratio scale to drift scale
}
