% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_ovrlbias.R
\name{ovrlbias}
\alias{ovrlbias}
\title{Overall bias of group sequential design}
\usage{
ovrlbias(t_vec, spd, alpha, mu)
}
\arguments{
\item{t_vec}{timing vector of interim looks and final analysis in information fraction. For example,
a trial design of one interim look at half information and final analysis has \code{t_vec=c(0.5,1)}}

\item{spd}{1 and 2 correspond to alpha spending functions which give O'Brien Fleming and Pocock
type boundaries, respectively. A value of 3 is the power family. See below for details.}

\item{alpha}{type I error, default is 0.05. See below for details.}

\item{mu}{drift parameter}
}
\value{
overall bias in drift scale.
}
\description{
Calculate the overall bias
}
\section{Details}{
 Argument \code{spd} and \code{alpha} essentially utilize arguments \code{iuse} and
\code{alpha} of function \code{bounds}	from package "\code{ldbounds}". Please refer to
package "\code{ldbounds}" for more details.
}

\examples{
#The overall bias for a trial with 4 interim looks at information
#fraction 0.2,0.4,0.6 and 0.8, with Pocock type boundary, and drift 3
> ovrlbias(t_vec = c(0.2,0.4,0.6,0.8,1),spd = 2,alpha = 0.05,mu = 3)
stop at: 1 
stop at: 2 
stop at: 3 
stop at: 4 
stop at: 5 
[1] 0.6308207
}
