% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_redensity.R
\name{redensity}
\alias{redensity}
\title{Recursive density function describing the probability of \eqn{(W(T)-\mu T; T)}}
\usage{
redensity(u, i, sigmas, cstars, bstars)
}
\arguments{
\item{u}{Input of function}

\item{i}{Index of interim look, e.g., \code{i=2} for the second interim look.
\code{i=length(t_vec)} for the final analysis.}

\item{sigmas}{Vector of \eqn{\sigmai}, same length as \code{cstars} and \code{bstars}.}

\item{cstars}{Vector of upper boundaries for \eqn{W(T)-\mu T}.}

\item{bstars}{Vector of lower boundaries for \eqn{W(T)-\mu T}.}
}
\value{

}
\description{
Recursive density function describing the probability of \eqn{(W(T)-\mu T; T)}
}
