% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_wtmut.R
\name{wtmut}
\alias{wtmut}
\title{Expectation of \eqn{(W(T)-\mu T; T=ti)}}
\usage{
wtmut(i, sigmas, cstars, bstars)
}
\arguments{
\item{i}{Index of interim look, e.g., \code{i=2} for the second interim look.
\code{i=length(t_vec)} for the final analysis.}

\item{sigmas}{Vector of \eqn{\sigmai}, same length as \code{cstars} and \code{bstars}.}

\item{cstars}{Vector of upper boundaries for \eqn{W(T)-\mu T}.}

\item{bstars}{Vector of lower boundaries for \eqn{W(T)-\mu T}.}
}
\value{
Expectation of \eqn{W(T)-\mu T; T=ti} depending on \code{i}.
}
\description{
Expectation of \eqn{(W(T)-\mu T; T=ti)} given \code{ti}
}
